requirements.txt
questionary
requests
deep-translator
langdetect
fastmcp
anyio
